Can talk to crush and get the api.
I had to modify parse_options() in postreview.py to gain access to the global options object. Not sure why and would rather not have to do that, but for now it gets me unstuck.

With that, I can create a PerforceClient object directly and talk to crush, so making progress.