Some refactoring and give user a second chance in new_change.
Since run_cmd was only called in the P4 class, I moved it there
and renamed it to run(). Then I renamed _p4_run() to run_G(). Those
names make more sense to me.

With that in place, I added code to new_change that gives the user
a second chance to edit the change list spec if it fails to load.