Determine whether to use old api or new when setting review status.
I was getting 500 errors on our 1.5.5 server when attempting to set
the changelist number and status for a submit. I'm trying to work around
it by going back to the old api code for setting status and skipping
the changelist update.  Note that the same api2 calls work fine in the
python post script, so I really don't know why it fails with the perl
script. However, if this works it's good enough for a deprecated script.