Better handling of edit --shelve
First it created an extra blank box on the review page. I made that go away but then it became a no-op.  This change handles it the same way post-review does by upload a diff, even though it's empty. That way the blank box has a link in it and doesn't draw unwanted attention.