rb2: add a diff command
The -n,--output-diff option wasn't properly hooked up. In attempting to fix that, I realized it didn't fit into the UI pattern we had established and was messy. Now we have a proper diff command which will diff the default change list, or an optional change list number provided by the user.

Need to implement this in post next.