Fix bugs in how editing and rid work
The change description was not being updated when we used the edit command. It
turns out that it conflicts with use of the -rid option, which I was always
passing assuming it would be harmless.  So now I have to only pass it when the
user passes it, which fixes the editing functionality.